enable misra-c2012-10.3

+++ board/drivers/fdcan.h
@@ -99,7 +99,7 @@ void process_can(uint8_t can_number) {
         if (can_check_checksum(&to_send)) {
           can_health[can_number].total_tx_cnt += 1U;
 
-          uint32_t TxFIFOSA = FDCAN_START_ADDRESS + (can_number * FDCAN_OFFSET) + (FDCAN_RX_FIFO_0_EL_CNT * FDCAN_RX_FIFO_0_EL_SIZE);
+          uint32_t TxFIFOSA = (uint32_t) (FDCAN_START_ADDRESS + (can_number * FDCAN_OFFSET) + (FDCAN_RX_FIFO_0_EL_CNT * FDCAN_RX_FIFO_0_EL_SIZE));
           // get the index of the next TX FIFO element (0 to FDCAN_TX_FIFO_EL_CNT - 1)
           uint32_t tx_index = (FDCANx->TXFQS >> FDCAN_TXFQS_TFQPI_Pos) & 0x1F;
           // only send if we have received a packet
@@ -168,7 +168,7 @@ void can_rx(uint8_t can_number) {
       can_health[can_number].total_rx_lost_cnt += 1U; // At least one message was lost
     }
 
-    uint32_t RxFIFO0SA = FDCAN_START_ADDRESS + (can_number * FDCAN_OFFSET);
+    uint32_t RxFIFO0SA = (uint32_t) (FDCAN_START_ADDRESS + (can_number * FDCAN_OFFSET));
     CANPacket_t to_push;
     canfd_fifo *fifo;
